[refactor] 디렉토리 이동 및 파일명 수정 #27
