+++ src/components/WeakApart/WholeWeakApartInfo/DetailWeakRankInfo.module.scss
