+++ src/components/WeakApart/WholeWeakApartInfo/WeakBuildRank.module.scss
