+++ src/components/WeakApart/WholeWeakApartInfo/WeakBuildRank.tsx
