+++ src/components/WeakApart/WholeWeakApartInfo/WholeWeakApartInfo.module.scss
