+++ src/components/WeakApart/WholeWeakApartInfo/WholeWeakApartInfo.tsx
@@ -1,7 +1,7 @@
 import styles from "./WholeWeakApartInfo.module.scss";
-import WeakMainNews from "./WeakMainNews/WeakMainNews";
+import WeakMainNews from "../WeakMainNews/WeakMainNews";
 import WeakBuildRank from "./WeakBuildRank";
-import ApartSearch from "../Common/ApartSearch";
+import ApartSearch from "../../Common/ApartSearch";
 
 export default function WholeWeakApartInfo() {
   return (
