+++ src/pages/WeakApartInfoPage.tsx
@@ -1,5 +1,5 @@
 import ApartMap from "@/components/Main/Map";
-import WholeWeakApartInfo from "@/components/WeakApart/WholeWeakApartInfo";
+import WholeWeakApartInfo from "@/components/WeakApart/WholeWeakApartInfo/WholeWeakApartInfo";
 import styels from "./WeakApartInfoPage.module.scss";
 
 export default function WeakApartInfoPage() {
